+++ typescript/zetachain/zetacore/observer/chain_nonces_pb.d.ts
@@ -16,7 +16,10 @@ export declare class ChainNonces extends Message<ChainNonces> {
   creator: string;
 
   /**
-   * @generated from field: string index = 2;
+   * deprecated(v19): index has been replaced by chain_id for unique identifier
+   *
+   * @generated from field: string index = 2 [deprecated = true];
+   * @deprecated
    */
   index: string;