+++ x/crosschain/client/cli/query_inbound_tracker.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"strconv"
 
+	"cosmossdk.io/errors"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/spf13/cobra"
@@ -11,6 +12,37 @@ import (
 	"github.com/zeta-chain/zetacore/x/crosschain/types"
 )
 
+func CmdShowInboundTracker() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "show-inbound-tracker [chainID] [txHash]",
+		Short: "shows an inbound tracker by chainID and txHash",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+			queryClient := types.NewQueryClient(clientCtx)
+			argChain, err := strconv.ParseInt(args[0], 10, 64)
+			if err != nil {
+				return errors.Wrapf(err, "unable to parse chain id from %q", args[0])
+			}
+			params := &types.QueryInboundTrackerRequest{
+				ChainId: argChain,
+				TxHash:  args[1],
+			}
+			res, err := queryClient.InboundTracker(context.Background(), params)
+			if err != nil {
+				return errors.Wrapf(
+					err,
+					"failed to fetch inbound tracker for chain %d and tx hash %s",
+					argChain,
+					args[1],
+				)
+			}
+			return clientCtx.PrintProto(res)
+		},
+	}
+	flags.AddQueryFlagsToCmd(cmd)
+	return cmd
+}
 func CmdListInboundTrackerByChain() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "list-inbound-tracker [chainId]",