+++ x/crosschain/keeper/msg_server_migrate_tss_funds_test.go
@@ -71,7 +71,6 @@ func setupTssMigrationParams(
 		MedianIndex:  1,
 	})
 	k.GetObserverKeeper().SetChainNonces(ctx, observertypes.ChainNonces{
-		Index:   chain.ChainName.String(),
 		ChainId: chain.ChainId,
 		Nonce:   1,
 	})