+++ x/observer/genesis_test.go
@@ -35,9 +35,9 @@ func TestGenesis(t *testing.T) {
 			LastObserverCount: sample.LastObserverCount(10),
 			TssFundMigrators:  []types.TssFundMigratorInfo{sample.TssFundsMigrator(1), sample.TssFundsMigrator(2)},
 			ChainNonces: []types.ChainNonces{
-				sample.ChainNonces(t, "0"),
-				sample.ChainNonces(t, "1"),
-				sample.ChainNonces(t, "2"),
+				sample.ChainNonces(0),
+				sample.ChainNonces(1),
+				sample.ChainNonces(2),
 			},
 			PendingNonces: sample.PendingNoncesList(t, "sample", 20),
 			NonceToCctx:   sample.NonceToCctxList(t, "sample", 20),