+++ x/observer/keeper/migrator.go
@@ -2,6 +2,8 @@ package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	v8 "github.com/zeta-chain/zetacore/x/observer/migrations/v8"
 )
 
 // Migrator is a struct for handling in-place store migrations.
@@ -41,4 +43,9 @@ func (m Migrator) Migrate5to6(_ sdk.Context) error {
 // Migrate6to7 migrates the store from consensus version 6 to 7
 func (m Migrator) Migrate6to7(_ sdk.Context) error {
 	return nil
+}
+
+// Migrate7to8 migrates the store from consensus version 7 to 8
+func (m Migrator) Migrate7to8(ctx sdk.Context) error {
+	return v8.MigrateStore(ctx, m.observerKeeper)
 }