+++ zetaclient/zetacore/client_worker.go
@@ -11,12 +11,12 @@ import (
 
 var logSampler = &zerolog.BasicSampler{N: 10}
 
-// UpdateZetacoreContextWorker is a polling goroutine that checks and updates zetacore context at every height.
+// UpdateAppContextWorker is a polling goroutine that checks and updates AppContext at every height.
 // todo implement graceful shutdown and work group
-func (c *Client) UpdateZetacoreContextWorker(ctx context.Context, app *appcontext.AppContext) {
+func (c *Client) UpdateAppContextWorker(ctx context.Context, app *appcontext.AppContext) {
 	defer func() {
 		if r := recover(); r != nil {
-			c.logger.Error().Interface("panic", r).Msg("UpdateZetacoreContextWorker: recovered from panic")
+			c.logger.Error().Interface("panic", r).Msg("UpdateAppContextWorker: recovered from panic")
 		}
 	}()
 
@@ -26,17 +26,17 @@ func (c *Client) UpdateZetacoreContextWorker(ctx context.Context, app *appcontex
 		logger      = c.logger.Sample(logSampler)
 	)
 
-	c.logger.Info().Msg("UpdateZetacoreContextWorker started")
+	c.logger.Info().Msg("UpdateAppContextWorker started")
 
 	for {
 		select {
 		case <-ticker.C:
-			c.logger.Debug().Msg("UpdateZetacoreContextWorker invocation")
-			if err := c.UpdateZetacoreContext(ctx, app, false, logger); err != nil {
-				c.logger.Err(err).Msg("UpdateZetacoreContextWorker failed to update config")
+			c.logger.Debug().Msg("UpdateAppContextWorker invocation")
+			if err := c.UpdateAppContext(ctx, app, false, logger); err != nil {
+				c.logger.Err(err).Msg("UpdateAppContextWorker failed to update config")
 			}
 		case <-c.stop:
-			c.logger.Info().Msg("UpdateZetacoreContextWorker stopped")
+			c.logger.Info().Msg("UpdateAppContextWorker stopped")
 			return
 		}
 	}